MVC아키텍쳐 적용
